Move singleton to key_value.hpp to reduce duplication and fix tests
